@@ -1,14 +1,11 @@
-apply plugin: 'android-sdk-manager'
 apply plugin: 'com.android.application'
 
-
 android {
-    compileSdkVersion 22
-    buildToolsVersion "22.0.1"
+    compileSdkVersion 28
 
     defaultConfig {
-        minSdkVersion 8
-        targetSdkVersion 22
+        minSdkVersion 14
+        targetSdkVersion 28
         versionCode 2
         versionName "1.0.1"
     }
@@ -20,9 +17,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':library')
-    compile 'com.android.support:appcompat-v7:22.+'
-    compile 'com.nineoldandroids:library:2.4.0'
-    compile 'com.daimajia.androidanimations:library:1.0.3@aar'
+    api fileTree(dir: 'libs', include: ['*.jar'])
+    api project(':library')
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'com.nineoldandroids:library:2.4.0'
+    implementation 'com.daimajia.androidanimations:library:1.0.3@aar'
 }
